Add Spew to dump entire Viper, plus AllSettingsLevels
viper.Spew() dumps the entire Viper object. There's a lot of
other debug code that will probably be deleted soon.

Added AllSettingsLevels which returns a map for each key that
declares where the key was found. Find a way to unify this with
AllSettings; the intention is for diagnostic output.